requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Pydantic v2 + settings
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0

# Database
sqlalchemy[asyncio]==2.0.38
asyncpg==0.30.0
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20

# HTTP client
httpx==0.28.1

# Embeddings — local (no API key needed)
sentence-transformers==3.4.1
numpy==2.2.2

# Vector DB support
pgvector==0.3.6

# Gemini — only used for venue text enrichment, not embeddings
google-genai==1.7.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
aiosqlite==0.21.0
